Update IPL and Secure Monitor Rev1.0.14 rev2
 [IPL]
 - Fix the comments in source codes.
 - Correspond to make the SDHI (DMAC) specification change.
 - QAC pointed out correspondence.
 - Fix flush process of the Console FIFO.
 - Add the System module stop control registers setting.
 - Modify version display of the M3 Ver2.0 to "M3 Ver1.1".
 - Update H3 v20 QoS setting rev.0.14
 - Update M3 v11 QoS setting rev.0.11
 - Delete debug code.
 - BL33x loadable range check does not work properly.

 [Secure Monitor]
 - Change module stop control from System to Secure.
 - Add the I-Cache clear.

+++ plat/renesas/rcar/include/bl2_dma_register.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015-2016, Renesas Electronics Corporation
+ * Copyright (c) 2015-2017, Renesas Electronics Corporation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -35,16 +35,16 @@
 #define	DMACH	0	/* DMA CH setting (0/16/32) */
 
 #if (DMACH==0)		/* SYS-DMAC0 (CH0) */
-#define SYS_DMAC_BIT	(1U << 19U)
+#define SYS_DMAC_BIT	((uint32_t)1U << 19U)
 #define DMA_BASE	(0xE6700000U)
 #elif (DMACH==16)	/* SYS-DMAC1 (CH16) */
-#define SYS_DMAC_BIT	(1U << 18U)
+#define SYS_DMAC_BIT	((uint32_t)1U << 18U)
 #define DMA_BASE	(0xE7300000U)
 #elif (DMACH==32)	/* SYS-DMAC2 (CH32) */
-#define SYS_DMAC_BIT	(1U << 17U)
+#define SYS_DMAC_BIT	((uint32_t)1U << 17U)
 #define DMA_BASE	(0xE7320000U)
 #else
-#define SYS_DMAC_BIT	(1U << 19U)
+#define SYS_DMAC_BIT	((uint32_t)1U << 19U)
 #define DMA_BASE	(0xE6700000U)
 #endif
